separated methods dealing with category names and its trees
