+++ lib/cyclopedio/mapping/candidate_generator.rb
@@ -4,6 +4,7 @@ require 'cyclopedio/syntax'
 require 'wiktionary/noun'
 require_relative 'name_mapper'
 require_relative 'candidate_set'
+require_relative 'wikipedia_category_utils'
 
 module Cyclopedio
   module Mapping
@@ -17,16 +18,16 @@ module Cyclopedio
         @cyc = options[:cyc] || Cyc::Client.new(cache: true)
         @name_mapper = options[:name_mapper] || NameMapper.new(cyc: @cyc, name_service: @name_service)
         @simplifier_factory = options[:simplifier_factory] || Cyclopedio::Syntax::Stanford::Simplifier
-        @parse_tree_factory = options[:parse_tree_factory] || Cyclopedio::Syntax::Stanford::Converter
         @candidate_set_factory = options[:candidate_set_factory] || CandidateSet
 
         @category_filters = options[:category_filters] || []
         @article_filters = options[:article_filters] || []
         @genus_filters = options[:genus_filters] || []
 
-        @nouns = options[:nouns] || Wiktionary::Noun.new
         @all_subtrees = options.fetch(:all_subtrees,false)
         @category_exact_match = options.fetch(:category_exact_match,false)
+
+        @wikipedia_category_utils = options[:wikipedia_category_utils] || WikipediaCategoryUtils.new
 
         @category_cache = Ref::WeakValueMap.new
         @article_cache = Ref::WeakValueMap.new
@@ -40,13 +41,13 @@ module Cyclopedio
         return @category_cache[category] unless @category_cache[category].nil?
         # from whole name singularized
         candidates = []
-        singularize_name(category.name, category_head(category)).each do |name_singularized|
+        @wikipedia_category_utils.singularize_name(category.name, @wikipedia_category_utils.category_head(category)).each do |name_singularized|
           candidates.concat(candidates_for_name(name_singularized,@category_filters))
         end
         candidate_set = create_candidate_set(category.name,candidates)
         return @category_cache[category] = candidate_set if !candidate_set.empty? || @category_exact_match
         # from simplified name
-        candidate_set = candidate_set_for_syntax_trees(category_head_trees(category),@category_filters)
+        candidate_set = candidate_set_for_syntax_trees(@wikipedia_category_utils.category_head_trees(category),@category_filters)
         return @category_cache[category] = candidate_set unless candidate_set.empty?
         # from original whole name
         candidate_set = candidate_set_for_name(category.name, @category_filters)
@@ -65,7 +66,7 @@ module Cyclopedio
         return @article_cache[article] unless @article_cache[article].nil?
         candidate_set = candidate_set_for_name(article.name, @article_filters)
         if candidate_set.empty?
-          candidate_set = candidate_set_for_name(remove_parentheses(article.name), @article_filters)
+          candidate_set = candidate_set_for_name(@wikipedia_category_utils.remove_parentheses(article.name), @article_filters)
         end
         @article_cache[article] = candidate_set
       end
@@ -82,7 +83,7 @@ module Cyclopedio
       # The result is a CandidateSet.
       def parentheses_candidates(concept)
         # TODO cache result for a given type
-        type = type_in_parentheses(concept.name)
+        type = @wikipedia_category_utils.type_in_parentheses(concept.name)
         candidate_set = create_candidate_set(type,[])
         if !type.empty?
           candidate_set = candidate_set_for_name(type, @genus_filters)
@@ -100,32 +101,6 @@ module Cyclopedio
       private
       # XXX most of these methods should be moved to external class(es).
 
-      # Returns the first parse tree of the +category+ name.
-      def category_head_tree(category)
-        convert_head_into_object(category.parsed_head)
-      end
-
-      # Returns the word (string) being a head of the +category+ name.
-      def category_head(category)
-        head_node = category_head_tree(category).find_head_noun
-        head_node && head_node.content
-      end
-
-      # Returns a list of parse trees of the +category+ name.
-      # The might be more parse trees for categories such as "Cities and
-      # villages in X", etc.
-      def category_head_trees(category)
-        if category.multiple_heads?
-          category.parsed_heads.map{|h| convert_head_into_object(h) }
-        else
-          [category_head_tree(category)]
-        end
-      end
-
-      # Converts the string representing the parsed +head+ into tree of objects.
-      def convert_head_into_object(head)
-        @parse_tree_factory.new(head).object_tree
-      end
 
       # Return candidate set for an entity +name+, with a given +head+ and apply
       # provided +filters+.
@@ -152,7 +127,7 @@ module Cyclopedio
           next unless head_node
           head = head_node.content
           names.each do |name|
-            simplified_names = singularize_name(name, head)
+            simplified_names = @wikipedia_category_utils.singularize_name(name, head)
             candidates = []
             simplified_names.each do |simplified_name|
               candidates.concat(candidates_for_name(simplified_name, filters))
@@ -173,29 +148,6 @@ module Cyclopedio
         filters.inject(candidates) do |terms, filter|
           filter.apply(terms)
         end
-      end
-
-      # Singularize using Wiktionary data. The result is an array of Strings.
-      def singularize_name(name, head)
-        names = []
-        singularized_heads = @nouns.singularize(head)
-        if singularized_heads
-          singularized_heads.each do |singularized_head|
-            names << name.sub(/\b#{Regexp.quote(head)}\b/, singularized_head)
-          end
-        end
-        names << name if names.empty?
-        names
-      end
-
-      def remove_parentheses(name)
-        return name if name !~ /\(/ || name =~ /^\(/
-        name.sub(/\([^)]*\)/,"").strip
-      end
-
-      def type_in_parentheses(name)
-        type = name[/\([^)]*\)/]
-        type ? type[1..-2] : ""
       end
 
       # Create a candidate set for single group of candidates.
